liveHTTP: fix signedness warnings

+++ modules/access_output/livehttp.c
@@ -500,7 +500,7 @@ static void destroySegment( output_segment_t *segment )
 static uint32_t segmentAmountNeeded( sout_access_out_sys_t *p_sys )
 {
     float duration = .0f;
-    for( unsigned index = 1; index <= vlc_array_count( p_sys->segments_t ) ; index++ )
+    for( unsigned index = 1; (int)index <= vlc_array_count( p_sys->segments_t ); index++ )
     {
         output_segment_t* segment = vlc_array_item_at_index( p_sys->segments_t, vlc_array_count( p_sys->segments_t ) - index );
         duration += segment->f_seglength;
@@ -525,7 +525,7 @@ static bool isFirstItemRemovable( sout_access_out_sys_t *p_sys, uint32_t i_first
     /* Check that segment has been out of playlist for seglenght + (p_sys->i_numsegs * p_sys->i_seglen) amount
      * We check this by calculating duration of the items that replaced first item in playlist
      */
-    for(int index=0; index < i_index_offset; index++ )
+    for( unsigned int index = 0; index < i_index_offset; index++ )
     {
         output_segment_t *segment = vlc_array_item_at_index( p_sys->segments_t, p_sys->i_segment - i_firstseg + index );
         duration += segment->f_seglength;
